shift method to local scope
